refactor: rename mongo script to old, to make way for new postgres
